+++ app/crud/__init__.py
@@ -1,3 +1,4 @@
 from .StepCRUD import stepCRUD
 from .UserCRUD import userCRUD
 from .RoleCRUD import roleCRUD
+from .DatasetCRUD import datasetCRUD